Stop merge_ranges claiming a sector it never covered, and cap both probe paths
The LOW/NIT tail, now in scope at the raised bar. 15 fixed, 2 re-rejected.

merge_ranges carried a +1 slack that treated half-open ranges as adjacent when
they merely touched at start == last_end, so [(0,5),(6,3)] came back as (0,9) —
claiming sector 5, which neither input covered. An existing test asserted the
wrong answer while its own comment said the opposite ("Absent: 1011"); the
slack was why. Test corrected, plus an explicit assertion that no range covers
the absent sector.

MAX_HDDVD_CLIPS and MAX_XPL_TITLES were both 4096, each giving a 64 GiB probe
product against an 8 GiB ceiling. Capping only the filed one would have moved
the amplification next door, so both are 512 and the test asserts the product
for each path — raising either constant fails.

Also: a .evo size accumulator that panicked on overflow from untrusted disc
content (both accumulators, one had been missed); Ok(0) reads now charge the
CSS crack budget rather than looping free; the VDS fallback dedupes on the
whole tuple so the wider sweep actually runs; host and .partial aliases no
longer let two entries both report complete; and disc-derived log fields use
the debug formatter so ANSI cannot reach a terminal verbatim.

Two findings re-rejected on evidence. ExtentNotUnitAligned's claimed zero-fill
is unreachable while fill_extents keeps its two is_source_terminated guards —
that rejection is conditional on those guards surviving. descramble_region's
Result<usize> is the decrypt_sectors family's documented always-zero legacy
count; changing one leaf would relocate the constant into a magic 0 and break a
public signature. The stale # Errors doc was fixed instead, and the tautological
out == 0 assertion moved up to where the contract actually lives.

+++ src/disc/mod.rs
@@ -3587,8 +3587,8 @@ mod tests {
     ) -> (crate::udf::fixture::MemDisc, udf::UdfFs) {
         use crate::udf::fixture::*;
         let files = vec![
-            file("MAIN.EVO", 100, 5_000, main_bytes, true),
-            file("OTHER.EVO", 101, 50_000, other_bytes, true),
+            file("MAIN.EVO", 100, 5_000, main_bytes as u64, true),
+            file("OTHER.EVO", 101, 50_000, other_bytes as u64, true),
         ];
         let root = DirSpec {
             name: String::new(),
